[common.h ssa.c]: Use long rather than int in runtime counters

+++ include/common.h
@@ -319,9 +319,9 @@ void ssa_twalk(const struct node_t *root,
 #endif
 
 void ssa_init_runtime_statistics();
-void ssa_set_runtime_counter(int id, int val);
-int  ssa_get_runtime_counter(int id);
-int  ssa_inc_runtime_counter(int id);
+void ssa_set_runtime_counter(int id, long val);
+long  ssa_get_runtime_counter(int id);
+long  ssa_inc_runtime_counter(int id);
 
 #ifdef __cplusplus
 }
